+++ deploy/complete/terraform/providers.tf
@@ -7,15 +7,15 @@ terraform {
   required_providers {
     oci = {
       source  = "hashicorp/oci"
-      version = ">= 4.23.0"
+      version = ">= 4.26.0"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = "1.11.2" # Latest version as March 2021 = 2.0.2. Using 1.11.2 (March, 2020) for ORM compatibility
+      version = "1.11.2" # Latest version as April 2021 = 2.1.0. Using 1.11.2 (March, 2020) for ORM compatibility
     }
     helm = {
       source  = "hashicorp/helm"
-      version = "1.1.1" # Latest version as March 2021 = 2.0.3. Using 1.1.1 (March, 2020) for ORM compatibility
+      version = "1.1.1" # Latest version as April 2021 = 2.1.1. Using 1.1.1 (March, 2020) for ORM compatibility
     }
     tls = {
       source  = "hashicorp/tls"
@@ -59,7 +59,6 @@ provider "oci" {
 
 # https://docs.cloud.oracle.com/en-us/iaas/Content/ContEng/Tasks/contengdownloadkubeconfigfile.htm#notes
 provider "kubernetes" {
-//  config_path = "/Users/pgressa/oracle/git/oraclecloud-cloudnative/deploy/complete/terraform/generated/kubeconfig"
   load_config_file       = "false" # Workaround for tf k8s provider < 1.11.1 to work with ORM
   cluster_ca_certificate = base64decode(yamldecode(data.oci_containerengine_cluster_kube_config.oke_cluster_kube_config.content)["clusters"][0]["cluster"]["certificate-authority-data"])
   host                   = yamldecode(data.oci_containerengine_cluster_kube_config.oke_cluster_kube_config.content)["clusters"][0]["cluster"]["server"]
@@ -79,7 +78,6 @@ provider "kubernetes" {
 # https://docs.cloud.oracle.com/en-us/iaas/Content/ContEng/Tasks/contengdownloadkubeconfigfile.htm#notes
 provider "helm" {
   kubernetes {
-//    config_path = "/Users/pgressa/oracle/git/oraclecloud-cloudnative/deploy/complete/terraform/generated/kubeconfig"
     load_config_file       = "false" # Workaround for tf helm provider < 1.1.1 to work with ORM
     cluster_ca_certificate = base64decode(yamldecode(data.oci_containerengine_cluster_kube_config.oke_cluster_kube_config.content)["clusters"][0]["cluster"]["certificate-authority-data"])
     host                   = yamldecode(data.oci_containerengine_cluster_kube_config.oke_cluster_kube_config.content)["clusters"][0]["cluster"]["server"]